refactor: move non-streamlit-dependent functions to lib
