+++ src/encord_active/lib/dataset/summary_utils.py
@@ -5,6 +5,17 @@ from typing import Tuple
 import numpy as np
 from encord.constants.enums import DataType
 
+from encord_active.lib.dataset.outliers import (
+    MetricOutlierInfo,
+    MetricsSeverity,
+    MetricWithDistanceSchema,
+    Severity,
+    get_iqr_outliers,
+)
+from encord_active.lib.metrics.utils import MetricData, load_metric_dataframe
+
+_COLUMNS = MetricWithDistanceSchema
+
 
 def get_all_image_sizes(project_folder: Path) -> np.ndarray:
     image_sizes = []
@@ -18,7 +29,7 @@ def get_all_image_sizes(project_folder: Path) -> np.ndarray:
     return np.array(image_sizes)
 
 
-def get_median_value_of_2D_array(array: np.ndarray) -> np.ndarray:
+def get_median_value_of_2d_array(array: np.ndarray) -> np.ndarray:
     """
     This function calculates the median value based on the product of the two dimension.
     For example, if they are image width and height, median dimensions corresponds to median average
@@ -46,3 +57,32 @@ def get_all_annotation_numbers(project_folder: Path) -> Tuple[int, int]:
                     classification_label_counter += len(data_unit["labels"]["classifications"])
 
     return classification_label_counter, object_label_counter
+
+
+def get_metric_summary(metrics: list[MetricData]) -> MetricsSeverity:
+    metric_severity = MetricsSeverity()
+    total_unique_severe_outliers = set()
+    total_unique_moderate_outliers = set()
+
+    for metric in metrics:
+        original_df = load_metric_dataframe(metric, normalize=False)
+        res = get_iqr_outliers(original_df)
+
+        if not res:
+            continue
+
+        df, iqr_outliers = res
+
+        df_dict = df.to_dict("records")
+        for row in df_dict:
+            if row[_COLUMNS.outliers_status] == Severity.severe:
+                total_unique_severe_outliers.add(row[_COLUMNS.identifier])
+            elif row[_COLUMNS.outliers_status] == Severity.moderate:
+                total_unique_moderate_outliers.add(row[_COLUMNS.identifier])
+
+        metric_severity.metrics.append(MetricOutlierInfo(metric=metric, df=df, iqr_outliers=iqr_outliers))
+
+    metric_severity.total_unique_severe_outliers = len(total_unique_severe_outliers)
+    metric_severity.total_unique_moderate_outliers = len(total_unique_moderate_outliers)
+
+    return metric_severity
